+++ Makefile
@@ -41,11 +41,11 @@ all:
 	
 	$(call printlog, BUILD, $(APPS1))
 	$(QUIET) (export PKG_CONFIG_PATH=$(PKG_CONFIG_PATH):$(ROOT_DIR)mk; \
-		$(GO) build -o $(APPS1) $(GOFLAGS) ./daemon/main.go)
+		$(GO) build -tags $(ENHANCED_KERNEL) -o $(APPS1) $(GOFLAGS) ./daemon/main.go)
 	
 	$(call printlog, BUILD, $(APPS2))
 	$(QUIET) (export PKG_CONFIG_PATH=$(PKG_CONFIG_PATH):$(ROOT_DIR)mk; \
-		$(GO) build -o $(APPS2) $(GOFLAGS) ./cmd/main.go)
+		$(GO) build -tags $(ENHANCED_KERNEL) -o $(APPS2) $(GOFLAGS) ./cmd/main.go)
 	
 	$(call printlog, BUILD, "kernel")
 	$(QUIET) make -C kernel/ko_src
