+++ kernel/ko_src/Makefile
@@ -3,7 +3,7 @@ DIRS := $(shell find $(CURRENT_PATH) -maxdepth 1 -type d)
 BASE_DIRS := $(basename $(patsubst $(CURRENT_PATH)/%, %, $(DIRS)))
 BASE_DIRS := $(filter-out $(CURRENT_PATH), $(BASE_DIRS))
 
-
+ifeq ($(ENHANCED_KERNEL), enhanced)
 all:
 	@for dir in ${BASE_DIRS}; do 	\
 		make -C $(CURRENT_PATH)/$$dir;	\
@@ -24,3 +24,13 @@ clean:
 	@for dir in ${BASE_DIRS}; do 	\
 		make -C $(CURRENT_PATH)/$$dir clean > /dev/null 2>&1||exit; done
 	@rm -rf ../ko/*
+else
+all:
+
+install:
+
+uninstall:
+
+clean:
+
+endif
